+++ tests/Fixtures/Photo.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AsimAli\Pinpoint\Tests\Fixtures;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Photo extends Model
+{
+    public $timestamps = false;
+
+    protected $guarded = [];
+}